Modified IRequest and Request, so that rating is double

+++ src/main/java/Entities/IRequest.java
@@ -22,7 +22,7 @@ public interface IRequest {
 
     IRequest getPrevious();
 
-    void setRating(int rating);
+    void setRating(double rating);
 
-    int getRating();
+    double getRating();
 }
